Rename / move `MultiParentCasperTestUtil` to `util.GenesisBuilder`
