+++ casper/src/test/scala/coop/rchain/casper/MultiParentCasperDeploySpec.scala
@@ -12,7 +12,7 @@ import org.scalatest.{FlatSpec, Inspectors, Matchers}
 
 class MultiParentCasperDeploySpec extends FlatSpec with Matchers with Inspectors {
 
-  import MultiParentCasperTestUtil._
+  import coop.rchain.casper.util.GenesisBuilder._
 
   implicit val timeEff = new LogicalTime[Effect]
