+++ casper/src/test/scala/coop/rchain/casper/MultiParentCasperRholangSpec.scala
@@ -13,8 +13,8 @@ import org.scalatest.{FlatSpec, Inspectors, Matchers}
 
 class MultiParentCasperRholangSpec extends FlatSpec with Matchers with Inspectors {
 
-  import MultiParentCasperTestUtil._
   import RSpaceUtil._
+  import coop.rchain.casper.util.GenesisBuilder._
 
   implicit val timeEff: LogicalTime[Effect] = new LogicalTime[Effect]
