+++ casper/src/test/scala/coop/rchain/casper/util/ProtoUtilTest.scala
@@ -28,7 +28,7 @@ class ProtoUtilTest extends FlatSpec with Matchers with GeneratorDrivenPropertyC
     }
   }
 
-  import MultiParentCasperTestUtil._
+  import GenesisBuilder._
 
   implicit val timeEff = new LogicalTime[Effect]
